+++ misc/cherry-pick-typeshed.py
@@ -14,13 +14,11 @@ import subprocess
 import sys
 import tempfile
 
-from mypy.util import safe
-
 
 def parse_commit_title(diff: str) -> str:
     m = re.search("\n    ([^ ].*)", diff)
     assert m is not None, "Could not parse diff"
-    return safe(m.group(1))
+    return m.group(1)
 
 
 def main() -> None: