+++ misc/incremental_checker.py
@@ -47,8 +47,6 @@ from argparse import ArgumentParser, Namespace, RawDescriptionHelpFormatter
 from typing import Any, Dict, Final
 from typing_extensions import TypeAlias as _TypeAlias
 
-from mypy.util import safe
-
 CACHE_PATH: Final = ".incremental_checker_cache.json"
 MYPY_REPO_URL: Final = "https://github.com/python/mypy.git"
 MYPY_TARGET_FILE: Final = "mypy"
@@ -171,7 +169,7 @@ def filter_daemon_stats(output: str) -> tuple[str, dict[str, Any]]:
     for line in lines:
         m = re.match(r"(\w+)\s+:\s+(.*)", line)
         if m:
-            key, value = [safe(g) for g in m.groups()]
+            key, value = [g for g in m.groups()]
             stats[key] = value
         else:
             output_lines.append(line)